move Anbox Cloud category to the bottom
Reasoning: this category is aimed at established users that will
go the extra mile to find these topics. The docs should cater more
to the casual reader who might get interested in trying this out
than to those that are just looking for the release notes.
(We'll also need to clean up the index page to have direct links
to all very relevant information, for example, the release notes.)

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>