move files to correspond to index
Change the file and folder names so that the directory structure
corresponds to the links defined in the index page.
This way, we can use the index page to find the topic IDs for
publishing.

Changed a few of the links in the index page as well because
the new links make more sense.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>