generate ams-configuration.md from a YAML file
Move the content of the tables in ams-configuration.md to a YAML
file.
In the future, this file can be generated automatically from the
code, or it could be reused to add the config options to the CLI
as well.

Add a script that uses a template file ams-configuration.tmpl.md
and the YAML file to generate the final Markdown file.

Signed-off-by: Ruth Fuchss <ruth.fuchss@canonical.com>